refactor: classify viewport dispatch, query
